FIX: 점수가 너무 낮으면 결과 등록 안하게 처리, 랭킹 데이터 반환 형식 수정

+++ src/main/java/com/example/speed_you_back/entity/Score.java
@@ -28,6 +28,6 @@ public class Score
     private Profile profile;
 
     @Column(nullable = false) private String game;
-    @Column(nullable = false) private double score;
+    @Column(nullable = false) private int score;
     @Column(nullable = false) private LocalDateTime created_at;
 }
